+++ Brokenhurst/MHSystems/app/src/main/java/com/mh/systems/brokenhurst/activites/DashboardActivity.java
@@ -16,7 +16,6 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.facebook.drawee.backends.pipeline.Fresco;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.google.gson.reflect.TypeToken;
@@ -118,7 +117,6 @@ public class DashboardActivity extends BaseActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        Fresco.initialize(DashboardActivity.this); //Initialize facebook Fresco for round profile pic.
         setContentView(R.layout.activity_dashboard);
 
         /**
@@ -424,8 +422,6 @@ public class DashboardActivity extends BaseActivity {
                     public void failure(RetrofitError error) {
                         //you can handle the errors here
                         Log.e(LOG_TAG, "RetrofitError : " + error);
-
-                        callWeatherService();
                     }
                 });
     }