+++ Brokenhurst/MHSystems/app/src/main/java/com/mh/systems/brokenhurst/activites/YourAccountActivity.java
@@ -6,6 +6,7 @@ import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v7.widget.PopupMenu;
 import android.support.v7.widget.Toolbar;
+import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -16,10 +17,14 @@ import android.widget.TextView;
 import com.mh.systems.brokenhurst.R;
 import com.mh.systems.brokenhurst.constants.ApplicationGlobal;
 import com.mh.systems.brokenhurst.fragments.FinanceFragment;
+import com.mh.systems.brokenhurst.fragments.HandicapFragment;
 import com.mh.systems.brokenhurst.fragments.MyAccountTabFragment;
+import com.mh.systems.brokenhurst.fragments.MyDetailsFragment;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
+
+import static com.mh.systems.brokenhurst.fragments.MyAccountTabFragment.iLastTabPosition;
 
 public class YourAccountActivity extends BaseActivity {
 
@@ -30,7 +35,7 @@ public class YourAccountActivity extends BaseActivity {
 
     private boolean isFilterOpen;
 
-    private int iTabPosition;
+    private static int iTabPosition;
 
     /*********************************
      * INSTANCES OF CLASSES
@@ -71,6 +76,8 @@ public class YourAccountActivity extends BaseActivity {
     //Pop Menu to show Categories of Course Diary.
     PopupMenu popupMenu;
 
+    private int iOpenTabPosition;
+
     Intent intent;
 
      /* -- INTERNET CONNECTION PARAMETERS -- */
@@ -84,47 +91,67 @@ public class YourAccountActivity extends BaseActivity {
                 @Override
                 public boolean onMenuItemClick(MenuItem item) {
 
-                    switch (item.getItemId()) {
-                        case R.id.item_today:
-                            ((FinanceFragment) getFragmentInstance()).updateFilterControl(0);
-                            break;
+                    iOpenTabPosition = iTabPosition;
 
-                        case R.id.item_a_week:
-                            ((FinanceFragment) getFragmentInstance()).updateFilterControl(1);
-                            break;
+                    if (isOnline(YourAccountActivity.this)) {
+                        switch (item.getItemId()) {
+                            case R.id.item_today:
+                                if (getFragmentInstance() instanceof FinanceFragment) {
+                                    ((FinanceFragment) getFragmentInstance()).updateFilterControl(0);
+                                }
+                                break;
 
-                        case R.id.item_one_month:
-                            ((FinanceFragment) getFragmentInstance()).updateFilterControl(2);
-                            break;
+                            case R.id.item_a_week:
+                                if (getFragmentInstance() instanceof FinanceFragment) {
+                                    ((FinanceFragment) getFragmentInstance()).updateFilterControl(1);
+                                }
+                                break;
 
-                        case R.id.item_three_months:
-                            ((FinanceFragment) getFragmentInstance()).updateFilterControl(3);
-                            break;
+                            case R.id.item_one_month:
+                                if (getFragmentInstance() instanceof FinanceFragment) {
+                                    ((FinanceFragment) getFragmentInstance()).updateFilterControl(2);
+                                }
+                                break;
 
-                        case R.id.item_six_months:
-                            ((FinanceFragment) getFragmentInstance()).updateFilterControl(4);
-                            break;
+                            case R.id.item_three_months:
+                                if (getFragmentInstance() instanceof FinanceFragment) {
+                                    ((FinanceFragment) getFragmentInstance()).updateFilterControl(3);
+                                }
+                                break;
 
-                        case R.id.item_a_year:
-                            ((FinanceFragment) getFragmentInstance()).updateFilterControl(5);
-                            break;
+                            case R.id.item_six_months:
+                                if (getFragmentInstance() instanceof FinanceFragment) {
+                                    ((FinanceFragment) getFragmentInstance()).updateFilterControl(4);
+                                }
+                                break;
 
-                        case R.id.item_from_start:
-                            ((FinanceFragment) getFragmentInstance()).updateFilterControl(6);
-                            break;
+                            case R.id.item_a_year:
+                                if (getFragmentInstance() instanceof FinanceFragment) {
+                                    ((FinanceFragment) getFragmentInstance()).updateFilterControl(5);
+                                }
+                                break;
 
-                        case R.id.item_toggle_mode:
-                            intent = new Intent(YourAccountActivity.this, EditToggleDetailActivity.class);
-                            startActivity(intent);
+                            case R.id.item_from_start:
+                                if (getFragmentInstance() instanceof FinanceFragment) {
+                                    ((FinanceFragment) getFragmentInstance()).updateFilterControl(6);
+                                }
+                                break;
 
-                            break;
+                            case R.id.item_toggle_mode:
+                                intent = new Intent(YourAccountActivity.this, EditToggleDetailActivity.class);
+                                startActivity(intent);
 
-                        case R.id.item_edit_mode:
-                            intent = new Intent(YourAccountActivity.this, EditDetailsActivity.class);
-                            startActivity(intent);
-                            break;
+                                break;
+
+                            case R.id.item_edit_mode:
+                                intent = new Intent(YourAccountActivity.this, EditDetailsActivity.class);
+                                startActivity(intent);
+                                break;
+                        }
+                        return true;
+                    } else {
+                        return false;
                     }
-                    return true;
                 }
             };
 
@@ -146,13 +173,14 @@ public class YourAccountActivity extends BaseActivity {
             getSupportActionBar().setTitle("Your Account");
         }
 
-        tfRobotoMedium = Typeface.createFromAsset(getAssets(), "fonts/Roboto-Medium.ttf");
+              tfRobotoMedium = Typeface.createFromAsset(getAssets(), "fonts/Roboto-Medium.ttf");
         //tvMyAccountTitle.setTypeface(tfRobotoMedium);
 
-        //Load Default fragment of COURSE DIARY.
-        updateFragment(new MyAccountTabFragment());
+        iOpenTabPosition = getIntent().getExtras().getInt("iTabPosition");
 
         initFianaceCategory();
+
+        //myAccountTabFragment = new MyAccountTabFragment();
 
         ivFilter.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -161,6 +189,13 @@ public class YourAccountActivity extends BaseActivity {
             }
         });
         popupMenu.setOnMenuItemClickListener(mCourseTypeListener);
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+
+        updateFragment(new MyAccountTabFragment(getiOpenTabPosition()));
     }
 
     @Override
@@ -192,42 +227,28 @@ public class YourAccountActivity extends BaseActivity {
          */
         popupMenu = new PopupMenu(this, ivFilter);
 
-        boolean isHandicapFeature = loadPreferenceBooleanValue(ApplicationGlobal.KEY_HANDICAP_FEATURE, false);
+        /**
+         * Step 2: Inflate the menu resource. Here the menu resource is
+         * defined in the res/menu project folder
+         */
+        switch (getWhichTab()) {
+            case 2:
+                ivFilter.setImageResource(R.mipmap.ic_event);
+                popupMenu.inflate(R.menu.finance_menu);
+                break;
 
-        if(isHandicapFeature){
-            /**
-             * Step 2: Inflate the menu resource. Here the menu resource is
-             * defined in the res/menu project folder
-             */
-            switch (getWhichTab()) {
-                case 2:
-                    ivFilter.setImageResource(R.mipmap.ic_event);
-                    popupMenu.inflate(R.menu.finance_menu);
-                    break;
-
-                case 0:
-                case 1:
-                    ivFilter.setImageResource(R.mipmap.ic_mode_edit);
-                    popupMenu.inflate(R.menu.my_details_menu);
-                    break;
-            }
-        }else{
-            /**
-             * Step 2: Inflate the menu resource. Here the menu resource is
-             * defined in the res/menu project folder
-             */
-            switch (getWhichTab()) {
-                case 1:
-                    ivFilter.setImageResource(R.mipmap.ic_event);
-                    popupMenu.inflate(R.menu.finance_menu);
-                    break;
-
-                case 0:
-                    ivFilter.setImageResource(R.mipmap.ic_mode_edit);
-                    popupMenu.inflate(R.menu.my_details_menu);
-                    break;
-            }
+            case 0:
+            case 1:
+                ivFilter.setImageResource(R.mipmap.ic_mode_edit);
+                popupMenu.inflate(R.menu.my_details_menu);
+                break;
         }
+
+       /* if (getWhichTab() instanceof FinanceFragment) {
+
+        } else {
+
+        }*/
 
         popupMenu.setOnMenuItemClickListener(mCourseTypeListener);
 
@@ -315,5 +336,13 @@ public class YourAccountActivity extends BaseActivity {
         ivFilter.setVisibility(iVisibleType);
 
         initFianaceCategory();
+    }
+
+    public int getiOpenTabPosition() {
+        return iOpenTabPosition;
+    }
+
+    public void setiOpenTabPosition(int iOpenTabPosition) {
+        this.iOpenTabPosition = iOpenTabPosition;
     }
 }