+++ Porters Park/MHSystems/app/src/main/java/com/mh/systems/porterspark/util/API/WebServiceMethods.java
@@ -376,5 +376,17 @@ public interface WebServiceMethods {
      */
     @POST("/webapi/api/ClubsApp")
     public void contactUs(@Body ContactUsAPI contactUsAPI, Callback<JsonObject> response);
+
+    /**
+     * Declaration of TOP UP prices list web service declaration.
+     * <p/>
+     * TYPE : POST
+     *
+     * @param aClientId     : Client ID.
+     * @param aMemberId     : Member ID.
+     * @param response      : Top Up Price list.
+     */
+    @GET("/api/ApifsiGateway/TopUps")
+    public void getTopUpPricesList(@Query("aClientId") String aClientId, @Query("aMemberId") String aMemberId, Callback<JsonObject> response);
 }