+++ MHS Demo/MHSystems/CHANGELOG.md
@@ -8,6 +8,9 @@ Initial release
 - Fixed: Club News: Implements Serializable to pass club news data on detail screen instead of passing value one by one.
 - New:  Top up Functionality.
 
-25th Jan, 2016
+25th Jan, 2017
 - UPDATE: 'Friends' info on FRIENDS tab display information on new screen instead of Alert Dialog.
 - UPDATE: REMOVE 'Contact Us' from Settings. Create new tab of 'Contact Us' along with MEMBERS and FRIENDS tabs. Also update 'CONTACT US' information.
+
+#####2nd Feb, 2017
+    - UPDATE: Replace 'SimpleDrawerView' with 'CircleImageView' for better performance.