+++ Porters Park/MHSystems/CHANGELOG.md
@@ -3,9 +3,12 @@
 - Fixed: Handicap graph date format 'MM/dd/yyyy' changed to 'dd/MM/yyyy'.
 - Fixed: Club News: Implements Serializable to pass club news data on detail screen instead of passing value one by one.
 
- #####25th Jan, 2016
-    - UPDATE: 'Friends' info on FRIENDS tab display information on new screen instead of Alert Dialog.
-    - UPDATE: REMOVE 'Contact Us' from Settings. Create new tab of 'Contact Us' along with MEMBERS and FRIENDS tabs. Also update 'CONTACT US' information.
+    #####25th Jan, 2017
+     - UPDATE: 'Friends' info on FRIENDS tab display information on new screen instead of Alert Dialog.
+     - UPDATE: REMOVE 'Contact Us' from Settings. Create new tab of 'Contact Us' along with MEMBERS and FRIENDS tabs. Also update 'CONTACT US' information.
+
+    #####2nd Feb, 2017
+     - UPDATE: Replace 'SimpleDrawerView' with 'CircleImageView' for better performance.
 
 ##Version 1.4.0 (2016-12-02)